requirements.txt
langgraph>=1.1.0
langchain>=1.2.0
langchain-openai>=1.1.0
langchain-core>=1.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
